Fixed multiple build issues, and improved file structure.
